+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the Wealthica Connect SDK across versions will be documen
 
 ## [Unreleased]
 
+## [0.0.11]
+- Fix passing params to transactions.getList()
+- Use institutions instead of institutionId for transactions.getList()
+- institutionId for transactions.getList() is deprecated now and will be removed in future releases
+- Add positions.getList() method
+
 ## [0.0.10]
 - Update basic example dependencies